% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/epi_final_size.R
\name{epi_final_size}
\alias{epi_final_size}
\title{Calculate final epidemic size}
\usage{
epi_final_size(r0 = 2, contact_matrix, demography_vector)
}
\arguments{
\item{r0}{Basic reproduction number. Default it 2}

\item{contact_matrix}{Social contact matrix. Entry mm_ij gives average number of contacts in group i reported by participants in group j}

\item{demography_vector}{Demography vector. Entry pp_i gives proportion of total population in group i (model will normalise if needed)}
}
\description{
This function calculates final epidemic size using SIR model for heterogeneously mixing population
}
\examples{
epi_final_size()
}
\keyword{epidemic}
\keyword{model}

